Add examples related to some major printing events
In particular, for the event 'PrintStarted' a script 'stampa_start.py' is provided, where
the option '-f' is used to provide the filename being printed from CLI.
The plan is to use this option in combination with the event hooks available in the
file 'config.yaml', where some placeholders can be used to access the event payload fields.
For more details, please have a look at https://docs.octoprint.org/master/events/index.html.

The content of the file 'config.yaml' related to the management of the events is provided as
example in the file 'example.config.yaml'.