Add log time capabilities and job time duration estimation to scripts in 'python/events'
When a print starts, the current machine time is printed in a file. If
the file already exists, it is overridden.
Then, the current machine time is printed whenever the printer enter a
pause or resume state.
Whenever the print job ends, the total job time duration is estimated by
analyzing the time log file.
The job time duration is computed from summing together all the time
intervals, as follows:
  total time += (inizio|riprendi) - (pausa|fine)
The estimation fails if the time log file is inconsistent. This occurs
whenever a pause is not resumed, or a job has started but not ended.